+++ src/main/java/com/seoultech/ecc/team/datamodel/TeamEntity.java
@@ -19,7 +19,7 @@ import java.util.List;
 public class TeamEntity extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long teamId;
+    private Integer teamId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "subject_id", nullable = false)
