+++ src/main/java/com/seoultech/ecc/team/repository/SubjectRepository.java
@@ -3,5 +3,5 @@ package com.seoultech.ecc.team.repository;
 import com.seoultech.ecc.team.datamodel.SubjectEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubjectRepository extends JpaRepository<SubjectEntity, Long> {
+public interface SubjectRepository extends JpaRepository<SubjectEntity, Integer> {
 }
